Fixed error from dividing an integer

+++ lib/MotorController/MotorController.cpp
@@ -20,10 +20,10 @@ void MotorController::moveDirection(MoveData movement){
   int speed = movement.speed;
   int rotation = movement.rotation;
 
-  int frontRightSpeed;
-  int backRightSpeed;
-  int backLeftSpeed;
-  int frontLeftSpeed;
+  double frontRightSpeed;
+  double backRightSpeed;
+  double backLeftSpeed;
+  double frontLeftSpeed;
 
   if(angle == -1)
   {
@@ -48,10 +48,10 @@ void MotorController::moveDirection(MoveData movement){
       a = a * multiplier;
     }
 
-    frontRightSpeed = -((int) a) + rotation;
-    backRightSpeed = -((int) b) + rotation;
-    backLeftSpeed = ((int) a) + rotation;
-    frontLeftSpeed = ((int) b) + rotation;
+    frontRightSpeed = -a + rotation;
+    backRightSpeed = -b + rotation;
+    backLeftSpeed = a + rotation;
+    frontLeftSpeed = b + rotation;
 
     multiplier = speed / max(abs(frontRightSpeed), max(abs(backRightSpeed), max(abs(backLeftSpeed), abs(frontLeftSpeed))));
